+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.9.6 - Unreleased
 
+### Added
+* new hashing 36-character base with alias (alphanum / abc123 / 36)
+
 
 ## Version 0.9.5 - Released 2021-02-05